+++ src/demux/ogg.rs
@@ -520,6 +520,7 @@ impl<R: Read + Seek> Demuxer for OggDemuxer<R> {
             title: None,
             artist: None,
             album: None,
+            metadata: std::collections::HashMap::new(),
         };
 
         let ret = info.clone();